first kinda working hacky version
