+++ core/Cargo.toml
@@ -24,7 +24,7 @@ build = "build.rs"
 maintenance = { status = "actively-developed" }
 
 [features]
-default 			= ["serde_support","slog/max_level_debug", "slog/release_max_level_info", "ser_id_64", "use_local_executor"]
+default 			= ["serde_support","slog/max_level_trace", "slog/release_max_level_info", "ser_id_64", "use_local_executor"]
 silent_logging 		= ["serde_support", "slog/max_level_info", "slog/release_max_level_error"]
 low_latency 		= ["executors/ws-no-park"]
 ser_id_64 			= []
@@ -67,7 +67,6 @@ core_affinity 					= {version = "0.5", optional = true}
 
 # Network-specific
 bytes 							= "0.5"
-#futures 						= "0.1"
 bitfields 						= "0.2"
 mio 							= {version = "0.7.0", features = ["tcp", "os-poll"]}
 iovec 							= "0.1.1" # Match MIOs Version
