+++ core/src/component/lifecycle.rs
